More 3d map opacity fixes

+++ bundles/net.tourbook/src/net/tourbook/map3/layer/tourtrack/TourPositionColors.java
@@ -1,14 +1,14 @@
 /*******************************************************************************
  * Copyright (C) 2005, 2014  Wolfgang Schramm and Contributors
- * 
+ *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation version 2 of the License.
- * 
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License along with
  * this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110, USA
@@ -78,7 +78,7 @@ class TourPositionColors implements Path.PositionColors {
 
 	/**
 	 * Update colors in the color provider.
-	 * 
+	 *
 	 * @param allTours
 	 */
 	public void updateColorProvider(final ArrayList<TourData> allTours) {
@@ -91,7 +91,7 @@ class TourPositionColors implements Path.PositionColors {
 		}
 	}
 
-	void updateColors(final double trackOpacity) {
+   void updateColors(final int trackOpacity) {
 
 		_awtColorCache.setTrackOpacity(trackOpacity);
 	}
